+++ src/envcaster/__init__.py
@@ -31,7 +31,7 @@ from envcaster.dotenv import (
     read_dotenv,
 )
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 
 __all__ = [
     "env",